requirements.txt
numpy
numpy-quaternion
matplotlib
